Fixed size_t conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@65031 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/test/objmgr/test_objmgr_data.cpp
@@ -319,8 +319,8 @@ bool CTestOM::Thread_Run(int idx)
         sel.SetOverlapType(sel.eOverlap_Intervals);
     }
     
-    int seq_count = 0, all_ids_count = 0;
-    int all_feat_count = 0, all_desc_count = 0;
+    size_t seq_count = 0, all_ids_count = 0;
+    size_t all_feat_count = 0, all_desc_count = 0;
     bool ok = true;
     typedef list<CConstRef<CSeq_entry> > TEntries;
     TEntries all_entries;
